pyproject.toml
[project]
name = "agentic-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[project.scripts]
run-pattern = "agentic_patterns.scripts.run_pattern:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

